requirements.txt
numpy==2.4.2
opencv_python==4.13.0.92
Pillow==12.1.1
streamlit==1.54.0
tensorflow==2.20.0
